Attempt defining Requirement in terms of GObject
This allows us to register custom GObject types within the Glib type system and define custom properties. These custom properties can then be easily synchronised to graphical elements (e.g. labels) with Glib::Binding and SYNC_CREATE/BIDIRECTIONAL.

This is clearly the way forwards, but it'll be a pain to convert all uses of projects and subsystems.